# Use map[string]struct{} for compression map in Len