--- a/dns_bench_test.go
+++ b/dns_bench_test.go
@@ -1,6 +1,7 @@
 package dns
 
 import (
+	"fmt"
 	"net"
 	"testing"
 )
@@ -60,6 +61,31 @@ func BenchmarkMsgLengthPack(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_, _ = msg.Pack()
+	}
+}
+
+func BenchmarkMsgLengthMassive(b *testing.B) {
+	makeMsg := func(question string, ans, ns, e []RR) *Msg {
+		msg := new(Msg)
+		msg.SetQuestion(Fqdn(question), TypeANY)
+		msg.Answer = append(msg.Answer, ans...)
+		msg.Ns = append(msg.Ns, ns...)
+		msg.Extra = append(msg.Extra, e...)
+		msg.Compress = true
+		return msg
+	}
+	const name1 = "12345678901234567890123456789012345.12345678.123."
+	rrMx := testRR(name1 + " 3600 IN MX 10 " + name1)
+	answer := []RR{rrMx, rrMx}
+	for i := 0; i < 128; i++ {
+		rrA := testRR(fmt.Sprintf("example%03d.something%03delse.org. 2311 IN A 127.0.0.1", i/32, i%32))
+		answer = append(answer, rrA)
+	}
+	answer = append(answer, rrMx, rrMx)
+	msg := makeMsg(name1, answer, nil, nil)
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		msg.Len()
 	}
 }