`map[string]int` requires 8 bytes per entry to store the unused position information.

~I couldn't get any clean benchmarks from this, but `map[string]struct{}` definitely uses less memory than `map[string]int`.~

This does unfortunately make the related test cases slightly weaker as they no longer test that the `Len` code is finding compression points in the correct spot.

Updates #709.

Edit: I was able to get some better benchmarks that show a very notable improvement in memory use.

```
$ benchstat {old,new}.bench
name                       old time/op    new time/op    delta
MsgLength-12                  329ns ± 0%     328ns ± 0%   -0.27%  (p=0.040 n=9+9)
MsgLengthNoCompression-12    9.37ns ± 1%    9.34ns ± 2%     ~     (p=0.099 n=10+10)
MsgLengthPack-12             2.26µs ± 9%    2.36µs ±23%     ~     (p=0.780 n=9+10)
MsgLengthMassive-12          63.1µs ± 8%    53.0µs ± 3%  -15.89%  (p=0.000 n=10+10)

name                       old alloc/op   new alloc/op   delta
MsgLength-12                  32.0B ± 0%     32.0B ± 0%     ~     (all equal)
MsgLengthNoCompression-12     0.00B          0.00B          ~     (all equal)
MsgLengthPack-12               896B ± 0%      896B ± 0%     ~     (all equal)
MsgLengthMassive-12          20.2kB ± 0%    14.8kB ± 0%  -26.74%  (p=0.000 n=8+9)

name                       old allocs/op  new allocs/op  delta
MsgLength-12                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
MsgLengthNoCompression-12      0.00           0.00          ~     (all equal)
MsgLengthPack-12               8.00 ± 0%      8.00 ± 0%     ~     (all equal)
MsgLengthMassive-12             138 ± 0%       138 ± 0%     ~     (all equal)
```