@tmthrgd 2018-11-27
These were going to create a merge conflict regardless of ordering. I've manually rebased.